proj: defined project structure
