+++ src/backend/Common/Class1.cs
@@ -1,0 +1,4 @@
+﻿ namespace Common
+{
+	public class Class1 { }
+}
