+++ src/backend/SecretsProvider/Startup.cs
@@ -1,0 +1,29 @@
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+
+namespace Ralfred
+{
+	public class Startup
+	{
+		public void ConfigureServices(IServiceCollection services) { }
+
+		public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+		{
+			if (env.IsDevelopment())
+			{
+				app.UseDeveloperExceptionPage();
+			}
+
+			app.UseRouting();
+
+			app.UseEndpoints(endpoints =>
+			{
+				endpoints.MapGet("/", async context => { await context.Response.WriteAsync("Hello World!"); });
+			});
+		}
+	}
+}
